refactor:#97 reservation 도메인 기능 통합 및 테스트 통합
