+++ src/main/java/team4/footwithme/resevation/api/ReservationApi.java
@@ -4,13 +4,16 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.domain.Sort;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.web.bind.annotation.*;
 import team4.footwithme.global.api.ApiResponse;
+import team4.footwithme.member.jwt.PrincipalDetails;
 import team4.footwithme.resevation.service.ReservationService;
 import team4.footwithme.resevation.service.response.ReservationsResponse;
+import team4.footwithme.resevation.service.response.ReservationInfoResponse;
+import team4.footwithme.resevation.service.response.ReservationInfoDetailsResponse;
+
+import java.util.List;
 
 @RequiredArgsConstructor
 @RestController
@@ -24,5 +27,27 @@ public class ReservationApi {
             @RequestParam Long reservationId,
             @RequestParam(defaultValue = "0", required = false) Integer page) {
         return ApiResponse.ok(reservationService.findReadyReservations(reservationId, page));
+    }
+
+    /**
+     * 팀 경기 일정 조회
+     */
+    @GetMapping("/{teamId}")
+    public ApiResponse<List<ReservationInfoResponse>> getTeamReservationInfo(@PathVariable Long teamId) {
+        return ApiResponse.ok(reservationService.getTeamReservationInfo(teamId));
+    }
+
+    /**
+     * 팀 경기 일정 상세 조회
+     */
+    @GetMapping("/details/{reservationId}")
+    public ApiResponse<ReservationInfoDetailsResponse> getTeamReservationInfoDetails(@PathVariable Long reservationId) {
+        return ApiResponse.ok(reservationService.getTeamReservationInfoDetails(reservationId));
+    }
+
+    @DeleteMapping("/{reservationId}")
+    public ApiResponse<Long> deleteReservation (@PathVariable(value = "reservationId") Long reservationId,
+                                                @AuthenticationPrincipal PrincipalDetails principalDetails){
+        return ApiResponse.ok(reservationService.deleteReservation(reservationId, principalDetails.getMember()));
     }
 }
