+++ src/main/java/team4/footwithme/resevation/repository/GameRepository.java
@@ -3,6 +3,7 @@ package team4.footwithme.resevation.repository;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
@@ -12,6 +13,7 @@ import team4.footwithme.resevation.domain.GameStatus;
 import team4.footwithme.resevation.domain.Reservation;
 import team4.footwithme.stadium.domain.Stadium;
 
+import java.util.List;
 import java.util.Optional;
 
 @Repository
@@ -25,6 +27,14 @@ public interface GameRepository extends JpaRepository<Game, Long>, CustomGlobalR
     @Query("SELECT g FROM Game g WHERE g.isDeleted = 'false' AND g.gameId = :id")
     Optional<Game> findActiveById(@Param("id") Long id);
 
+    @Modifying
     @Query("UPDATE Game g SET g.isDeleted = 'TRUE' WHERE g.secondTeamReservation = :reservation")
     void softDeleteBySecondTeamReservation(@Param("reservation") Reservation reservation);
+
+    @Query("select g.firstTeamReservation from Game g where g.isDeleted = 'false' and g.secondTeamReservation.reservationId = :reservationId")
+    Optional<Reservation> findFirstTeamReservationBySecondTeamReservationId(@Param("reservationId") Long secondTeamReservationId);
+
+    @Query("select g from Game g where g.isDeleted = 'false' and g.firstTeamReservation.reservationId = :reservationId")
+    List<Game> findAllByReservationId(Long reservationId);
+
 }
