+++ src/main/java/team4/footwithme/resevation/service/request/MercenaryServiceRequest.java
@@ -1,6 +1,6 @@
 package team4.footwithme.resevation.service.request;
 
-public record MWMercenaryServiceRequest(
+public record MercenaryServiceRequest(
         Long reservationId,
         String description
 ) {
