+++ src/main/java/team4/footwithme/resevation/service/request/ParticipantUpdateServiceRequest.java
@@ -2,7 +2,7 @@ package team4.footwithme.resevation.service.request;
 
 import team4.footwithme.resevation.domain.ParticipantRole;
 
-public record MWParticipantUpdateServiceRequest(
+public record ParticipantUpdateServiceRequest(
         Long participantId,
         ParticipantRole role
 ) {
