+++ src/main/java/team4/footwithme/resevation/service/response/MercenaryResponse.java
@@ -2,11 +2,11 @@ package team4.footwithme.resevation.service.response;
 
 import team4.footwithme.resevation.domain.Mercenary;
 
-public record MWMercenaryResponse(
+public record MercenaryResponse(
         Long mercenaryId,
         Long reservationId,
         String description) {
-    public MWMercenaryResponse(Mercenary mercenary) {
+    public MercenaryResponse(Mercenary mercenary) {
         this(
                 mercenary.getMercenaryId(),
                 mercenary.getReservation().getReservationId(),
