+++ src/main/java/team4/footwithme/resevation/service/response/ParticipantInfoResponse.java
@@ -3,12 +3,13 @@ package team4.footwithme.resevation.service.response;
 import team4.footwithme.resevation.domain.Participant;
 import team4.footwithme.resevation.domain.ParticipantRole;
 
-public record YYParticipantResponse(
+// TODO :: ParticipantInfoResponse 로 변경 부탁
+public record ParticipantInfoResponse(
         String memberName,
         ParticipantRole role
 ) {
-    public static YYParticipantResponse from(Participant participant) {
-        return new YYParticipantResponse(
+    public static ParticipantInfoResponse from(Participant participant) {
+        return new ParticipantInfoResponse(
             participant.getMember().getName(),
             participant.getParticipantRole()
         );
