+++ src/main/java/team4/footwithme/resevation/service/response/ParticipantMemberInfo.java
@@ -3,13 +3,13 @@ package team4.footwithme.resevation.service.response;
 import team4.footwithme.member.domain.Member;
 import team4.footwithme.member.domain.MemberRole;
 
-public record MWParticipantMemberInfo(
+public record ParticipantMemberInfo(
         Long memberId,
         String email,
         String name,
         MemberRole memberRole) {
 
-    public MWParticipantMemberInfo(Member member) {
+    public ParticipantMemberInfo(Member member) {
         this(
                 member.getMemberId(),
                 member.getEmail(),
