+++ src/test/java/team4/footwithme/ApiTestSupport.java
@@ -17,13 +17,13 @@ import team4.footwithme.member.jwt.JwtTokenFilter;
 import team4.footwithme.member.service.CookieService;
 import team4.footwithme.member.service.MemberService;
 import team4.footwithme.resevation.api.GameApi;
+import team4.footwithme.resevation.api.MercenaryApi;
+import team4.footwithme.resevation.api.ParticipantApi;
 import team4.footwithme.resevation.api.ReservationApi;
-import team4.footwithme.resevation.api.MWMercenaryApi;
-import team4.footwithme.resevation.api.MWParticipantApi;
 import team4.footwithme.resevation.service.GameService;
+import team4.footwithme.resevation.service.MercenaryService;
+import team4.footwithme.resevation.service.ParticipantService;
 import team4.footwithme.resevation.service.ReservationService;
-import team4.footwithme.resevation.service.MWParticipantServiceImpl;
-import team4.footwithme.resevation.service.MercenaryServiceImpl;
 import team4.footwithme.stadium.api.CourtMerchantApi;
 import team4.footwithme.stadium.api.StadiumMerchantApi;
 import team4.footwithme.stadium.service.CourtService;
@@ -32,7 +32,7 @@ import team4.footwithme.vote.api.VoteApi;
 import team4.footwithme.vote.service.VoteService;
 
 @WebMvcTest(controllers = {
-    VoteApi.class, ChatApi.class, ChatroomApi.class, ChatMemberApi.class, StadiumMerchantApi.class, CourtMerchantApi.class, GameApi.class, ReservationApi.class, MemberApi.class, MWMercenaryApi.class, MWParticipantApi.class
+    VoteApi.class, ChatApi.class, ChatroomApi.class, ChatMemberApi.class, StadiumMerchantApi.class, CourtMerchantApi.class, GameApi.class, ReservationApi.class, MemberApi.class, MercenaryApi.class, ParticipantApi.class
 })
 @AutoConfigureMockMvc(addFilters = false)
 public abstract class ApiTestSupport {
@@ -56,10 +56,10 @@ public abstract class ApiTestSupport {
     protected ChatMemberService chatMemberService;
 
     @MockBean
-    protected MercenaryServiceImpl mercenaryService;
+    protected MercenaryService mercenaryService;
 
     @MockBean
-    protected MWParticipantServiceImpl participantService;
+    protected ParticipantService participantService;
 
     @MockBean
     protected JwtTokenFilter jwtTokenFilter;
