+++ src/test/java/team4/footwithme/docs/reservation/ReservationApiDocs.java
@@ -7,14 +7,19 @@ import org.springframework.http.MediaType;
 import org.springframework.restdocs.payload.JsonFieldType;
 import team4.footwithme.docs.RestDocsSupport;
 import team4.footwithme.resevation.api.ReservationApi;
+import team4.footwithme.resevation.domain.ParticipantGender;
+import team4.footwithme.resevation.domain.ParticipantRole;
+import team4.footwithme.resevation.domain.ReservationStatus;
 import team4.footwithme.resevation.service.ReservationService;
-import team4.footwithme.resevation.service.response.ReservationsResponse;
+import team4.footwithme.resevation.service.response.*;
 import team4.footwithme.security.WithMockPrincipalDetail;
 
 import java.time.LocalDateTime;
 import java.util.Collections;
+import java.util.List;
 
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mock;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
@@ -22,8 +27,8 @@ import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuild
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.*;
 import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath;
 import static org.springframework.restdocs.payload.PayloadDocumentation.responseFields;
-import static org.springframework.restdocs.request.RequestDocumentation.parameterWithName;
-import static org.springframework.restdocs.request.RequestDocumentation.queryParameters;
+import static org.springframework.restdocs.request.RequestDocumentation.*;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
@@ -90,6 +95,125 @@ public class ReservationApiDocs extends RestDocsSupport {
                                 fieldWithPath("data.sort.unsorted").type(JsonFieldType.BOOLEAN).description("정렬되지 않았는지 여부"),
                                 fieldWithPath("data.first").type(JsonFieldType.BOOLEAN).description("첫 번째 페이지 여부"),
                                 fieldWithPath("data.empty").type(JsonFieldType.BOOLEAN).description("데이터가 비어있는지 여부")
+                        )
+                ));
+    }
+
+    @DisplayName("팀 일정 전체조회 API")
+    @Test
+    void getTeamReservationInfo() throws Exception {
+        Long teamId = 1L;
+
+        List<ReservationInfoResponse> response = List.of(
+                new ReservationInfoResponse(
+                        "구장1",
+                        LocalDateTime.of(2024, 11,12,15,0,0),
+                        ReservationStatus.RECRUITING
+                ),
+                new ReservationInfoResponse(
+                        "구장2",
+                        LocalDateTime.of(2024, 11,20,13,0,0),
+                        ReservationStatus.RECRUITING
+                )
+        );
+
+        given(reservationService.getTeamReservationInfo(teamId))
+                .willReturn(response);
+
+        mockMvc.perform(get("/api/v1/reservation/{teamId}", teamId)
+                        .contentType(MediaType.APPLICATION_JSON))
+                .andExpect(status().isOk())
+                .andDo(document("reservation-info-get",
+                        preprocessRequest(prettyPrint()),
+                        preprocessResponse(prettyPrint()),
+                        pathParameters(
+                                parameterWithName("teamId").description("팀 ID")
+                        ),
+                        responseFields(
+                                fieldWithPath("code").type(JsonFieldType.NUMBER).description("응답 코드"),
+                                fieldWithPath("status").type(JsonFieldType.STRING).description("응답 상태"),
+                                fieldWithPath("message").type(JsonFieldType.STRING).description("응답 메시지"),
+                                fieldWithPath("data").type(JsonFieldType.ARRAY).description("팀 일정 리스트"),
+                                fieldWithPath("data[].courtName").type(JsonFieldType.STRING).description("구장 이름"),
+                                fieldWithPath("data[].matchDate").type(JsonFieldType.ARRAY).description("예약 날짜"),
+                                fieldWithPath("data[].status").type(JsonFieldType.STRING).description("진행 상태")
+                        )
+                ));
+    }
+
+    @DisplayName("팀 일정 상세 조회 API")
+    @Test
+    void getTeamReservationInfoDetails() throws Exception {
+        long reservationId = 1L;
+
+        ReservationInfoDetailsResponse response = new ReservationInfoDetailsResponse(
+                "구장 1",
+                "상대 팀",
+                LocalDateTime.of(2024, 11,12,15,0,0),
+                List.of(
+                        new ParticipantInfoResponse("팀원1", ParticipantRole.MEMBER),
+                        new ParticipantInfoResponse("팀원2", ParticipantRole.MEMBER)
+                ),
+                ParticipantGender.MALE,
+                ReservationStatus.CONFIRMED
+        );
+
+        given(reservationService.getTeamReservationInfoDetails(reservationId))
+                .willReturn(response);
+
+        mockMvc.perform(get("/api/v1/reservation/details/{reservationId}", reservationId)
+                        .contentType(MediaType.APPLICATION_JSON))
+                .andExpect(status().isOk())
+                .andDo(document("reservation-detailInfo-get",
+                        preprocessRequest(prettyPrint()),
+                        preprocessResponse(prettyPrint()),
+                        pathParameters(
+                                parameterWithName("reservationId").description("예약 ID")
+                        ),
+                        responseFields(
+                                fieldWithPath("code").type(JsonFieldType.NUMBER).description("응답 코드"),
+                                fieldWithPath("status").type(JsonFieldType.STRING).description("응답 상태"),
+                                fieldWithPath("message").type(JsonFieldType.STRING).description("응답 메시지"),
+                                fieldWithPath("data").type(JsonFieldType.OBJECT).description("응답 데이터"),
+                                fieldWithPath("data.courtName").type(JsonFieldType.STRING).description("구장 이름"),
+                                fieldWithPath("data.matchTeamName").type(JsonFieldType.STRING).description("상대팀 이름"),
+                                fieldWithPath("data.matchDate").type(JsonFieldType.ARRAY).description("예약 날짜"),
+                                fieldWithPath("data.participants").type(JsonFieldType.ARRAY).description("참여자 리스트"),
+                                fieldWithPath("data.participants[].memberName").type(JsonFieldType.STRING).description("참여자 이름"),
+                                fieldWithPath("data.participants[].role").type(JsonFieldType.STRING).description("참여자 상태"),
+                                fieldWithPath("data.gender").type(JsonFieldType.STRING).description("참여인원 성비"),
+                                fieldWithPath("data.status").type(JsonFieldType.STRING).description("진행 상태")
+                        )
+                ));
+    }
+
+    @DisplayName("예약 취소하는 API")
+    @Test
+    void deleteReservation() throws Exception {
+        //given
+        Long reservationId = 1L;
+
+        //when
+        given(reservationService.deleteReservation(eq(reservationId), any()))
+                .willReturn(reservationId);
+
+        //then
+        mockMvc.perform(delete("/api/v1/reservation/{reservationId}", reservationId)
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .header("Authorization", "ACCESS_TOKEN"))
+                .andExpect(status().isOk())
+                .andDo(document("reservation-delete",
+                        preprocessRequest(prettyPrint()),
+                        preprocessResponse(prettyPrint()),
+                        responseFields(
+                                fieldWithPath("code").type(JsonFieldType.NUMBER)
+                                        .description("코드"),
+                                fieldWithPath("status").type(JsonFieldType.STRING)
+                                        .description("상태"),
+                                fieldWithPath("message").type(JsonFieldType.STRING)
+                                        .description("메시지"),
+                                fieldWithPath("data").type(JsonFieldType.NUMBER)
+                                        .description("삭제된 예약 ID")
                         )
                 ));
     }
